refactor: Move setAttributeNS tests to elements.test.js
